# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apple Connector API",
    "description": "Hybrid HTTP API for Messages.app, Reminders.app, Notes.app, Calendar.app, and Contacts.\n\n**Reads** use live SQLite connections (requires Full Disk Access).\n\n**Writes** for Reminders, Calendar events, and Contacts use EventKit / Contacts framework (requires Reminders, Calendars, and Contacts permissions in System Settings). Unsupported fields return `422`; smart lists, read-only calendars/containers return `403`. Post-save responses hydrate from SQLite with up to 5×100ms retries; `sync_pending: true` indicates the SQLite read path has not caught up yet.\n\nPagination uses keyset cursors only (default limit 50, maximum 200, newest first). Offsets are not supported.\n\nAuthentication, TLS, and network exposure controls are expected to be enforced by an external reverse proxy or firewall. This service does not implement authentication or TLS.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this service (use the host you opened in the browser, e.g. http://127.0.0.1:3000 or http://localhost:3000)"
    }
  ],
  "paths": {
...
    "/v1/containers": {
      "get": {
        "tags": [
          "containers"
        ],
        "summary": "List contact containers",
        "operationId": "listContainers",
        "responses": {
          "200": {
            "description": "Contact containers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerPageDto"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/containers/{container_id}": {
      "get": {
        "tags": [
          "containers"
        ],
        "summary": "Get a contact container",
        "operationId": "getContainer",
        "parameters": [
          {
            "name": "container_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Container detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerSummaryDto"
                }
              }
            }
          },
          "404": {
            "description": "Container not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContainerId": {
        "type": "string",
        "description": "Stable identifier for a contact container (its UUID).",
        "example": "11111111-1111-1111-1111-111111111111"
      },
...
      "ContainerPageDto": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContainerSummaryDto"
            }
          }
        }
      },
      "ContainerSummaryDto": {
        "type": "object",
        "required": [
          "id",
          "source_id",
          "container_type",
          "read_only"
        ],
        "properties": {
          "container_type": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "$ref": "#/components/schemas/ContainerId"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "read_only": {
            "type": "boolean"
          },
          "source_id": {
            "$ref": "#/components/schemas/SourceId"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "validation_error",
          "not_found",
          "range_not_satisfiable",
          "service_unavailable",
          "forbidden",
          "conflict",
          "internal_error"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "SourceId": {
        "type": "string",
        "description": "AddressBook source UUID (directory name under Sources/).",
        "example": "27fd6c1e-5da5-4340-a31a-1d83c25d3b70"
      },
...
    },
    "securitySchemes": {
      "reverse_proxy": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentication is expected to be enforced by an external reverse proxy or firewall. This API does not implement authentication or TLS."
      }
    }
  },
  "security": [
    {
      "reverse_proxy": []
    }
  ],
...
}